+++ packages/date/src/Date.js
@@ -2,9 +2,8 @@ import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { useField, useFormikContext } from 'formik';
-import 'react-dates/initialize';
-import { SingleDatePicker } from 'react-dates';
-import Icon from '@availity/icon';
+import '@availity/react-dates/initialize';
+import { SingleDatePicker } from '@availity/react-dates';
 import { InputGroup, Input, Row, Col } from 'reactstrap';
 import moment from 'moment';
 import '../polyfills';
@@ -21,13 +20,11 @@ const yearPickerStyles = {
 const AvDate = ({
   className,
   name,
-  calendarIcon,
   innerRef,
   onChange,
   onPickerFocusChange,
   min,
   max,
-  datepicker,
   format,
   validate,
   datePickerProps,
@@ -45,8 +42,7 @@ const AvDate = ({
     className,
     metadata.touched ? 'is-touched' : 'is-untouched',
     metadata.touched && metadata.error && 'is-invalid',
-    !field.value && 'current-day-highlight',
-    datepicker && 'av-calendar-show'
+    !field.value && 'current-day-highlight'
   );
 
   const pickerId = `${(attributes.id || name).replace(/[^\da-z]/gi, '')}-picker`;
@@ -181,11 +177,9 @@ const AvDate = ({
           onFocusChange={onFocusChange}
           numberOfMonths={1}
           isOutsideRange={isOutsideRange(min, max)}
-          customInputIcon={datepicker ? calendarIcon : undefined}
-          showDefaultInputIcon={datepicker}
-          inputIconPosition="after"
           navPosition="navPositionBottom"
           openDirection={openDirection}
+          autoComplete="date"
         />
       </InputGroup>
     </>
@@ -198,22 +192,18 @@ AvDate.propTypes = {
   className: PropTypes.string,
   min: limitPropType,
   max: limitPropType,
-  calendarIcon: PropTypes.node,
   onChange: PropTypes.func,
   onPickerFocusChange: PropTypes.func,
   innerRef: PropTypes.oneOfType([PropTypes.func, PropTypes.node]),
   format: PropTypes.string,
   'data-testid': PropTypes.string,
-  datepicker: PropTypes.bool,
   validate: PropTypes.func,
   datePickerProps: PropTypes.object,
   openDirection: PropTypes.string,
 };
 
 AvDate.defaultProps = {
-  calendarIcon: <Icon name="calendar" />,
   format: 'MM/DD/YYYY',
-  datepicker: true,
   openDirection: 'down',
 };