+++ packages/feedback/src/FeedbackForm.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { Button, ModalBody, ModalHeader, ModalFooter, FormGroup } from 'reactstrap';
 import { avLogMessagesApi, avRegionsApi } from '@availity/api-axios';
@@ -40,11 +40,12 @@ const FeedbackForm = ({
   showSupport,
   setSupportIsActive,
   autoFocusFeedbackButton,
+  modal,
   ...formProps
 }) => {
   const [active, setActive] = useState(null);
   const [sent, setSent] = useState(null);
-
+  const ref = useRef();
   const sendFeedback = async ({ smileField, ...values }) => {
     const response = await avRegionsApi.getCurrentRegion();
 
@@ -110,10 +111,11 @@ const FeedbackForm = ({
         {prompt || `Tell us what you think about ${name}`}
       </ModalHeader>
       <Form
+        innerRef={ref}
         aria-label="Feedback Form"
         aria-describedby="feedback-form-header"
         role="form"
-        onKeyDown={({ keyCode }) => keyCode === 27 && onClose()}
+        onKeyDown={({ key }) => key === 'Escape' && onClose()}
         data-testid="feedback-form"
         initialValues={{
           'face-options': undefined,
@@ -151,7 +153,9 @@ const FeedbackForm = ({
               options={faceOptions}
               name="smileField"
               onChange={(option) => setActive(option)}
+              onClose={onClose}
               autoFocusFeedbackButton={autoFocusFeedbackButton}
+              modal={modal}
             />
           </FormGroup>
 
@@ -198,7 +202,7 @@ const FeedbackForm = ({
                 onClick={() => setSupportIsActive(true)}
                 color="link"
                 type="button"
-                onKeyDown={({ keyCode }) => keyCode === 13 && setSupportIsActive(true)}
+                onKeyDown={({ key }) => key === 'Enter' && setSupportIsActive(true)}
               >
                 Open a support ticket
               </Button>
@@ -206,12 +210,36 @@ const FeedbackForm = ({
           ) : null}
 
           {onClose ? (
-            <Button onClick={onClose} color="secondary" onKeyDown={({ keyCode }) => keyCode === 13 && onClose()}>
+            <Button
+              type="button"
+              onClick={onClose}
+              color="secondary"
+              onKeyDown={({ key, shiftKey }) => {
+                if (key === 'Enter') {
+                  onClose();
+                }
+                if (key === 'Tab' && !active && !shiftKey && !modal) {
+                  onClose();
+                }
+              }}
+            >
               Close
             </Button>
           ) : null}
 
-          <Button type="submit" color="primary" disabled={!active}>
+          <Button
+            onKeyDown={({ key, shiftKey }) => {
+              if (key === 'Enter') {
+                ref.current.submitForm();
+              }
+              if (key === 'Tab' && !shiftKey && !modal) {
+                onClose();
+              }
+            }}
+            type="submit"
+            color="primary"
+            disabled={!active}
+          >
             Send Feedback
           </Button>
         </ModalFooter>
@@ -248,6 +276,7 @@ FeedbackForm.propTypes = {
   showSupport: PropTypes.bool,
   setSupportIsActive: PropTypes.func,
   autoFocusFeedbackButton: PropTypes.bool,
+  modal: PropTypes.bool,
 };
 
 FeedbackForm.defaultProps = {
@@ -256,6 +285,7 @@ FeedbackForm.defaultProps = {
   additionalComments: false,
   analytics: avLogMessagesApi,
   showSupport: false,
+  modal: true,
 };
 
 export default FeedbackForm;