+++ packages/reactstrap-validation-date/src/AvDateRange.js
@@ -5,10 +5,9 @@ import moment from 'moment';
 import pick from 'lodash/pick';
 import { inputType, isoDateFormat } from 'availity-reactstrap-validation/lib/AvValidator/utils';
 import { AvInput } from 'availity-reactstrap-validation';
-import { DateRangePicker } from 'react-dates';
-import 'react-dates/initialize';
+import { DateRangePicker } from '@availity/react-dates';
+import '@availity/react-dates/initialize';
 import classNames from 'classnames';
-import Icon from '@availity/icon';
 
 import { isOutsideRange, limitPropType, isSameDay } from './utils';
 import '../polyfills';
@@ -425,7 +424,7 @@ class AvDateRange extends Component {
   };
 
   render() {
-    const { name, className, id, min, max, calendarIcon, datepicker, validate, distance, ...attributes } = this.props;
+    const { name, className, id, min, max, validate, distance, ...attributes } = this.props;
     const { startValue, endValue, focusedInput, format } = this.state;
     const endValidate = {
       afterStart: this.afterStartValidate,
@@ -468,8 +467,7 @@ class AvDateRange extends Component {
       isBad ? 'is-bad-input' : null,
       hasError ? 'av-invalid' : 'av-valid',
       validation.error && 'is-invalid',
-      !startValue && !endValue && 'current-day-highlight',
-      datepicker && 'av-calendar-show'
+      !startValue && !endValue && 'current-day-highlight'
     );
 
     return (
@@ -527,25 +525,11 @@ class AvDateRange extends Component {
             focusedInput={focusedInput}
             onFocusChange={this.onFocusChange}
             isOutsideRange={isOutsideRange(minDate, maxDate, format)}
-            customInputIcon={
-              datepicker
-                ? React.cloneElement(calendarIcon, {
-                    ref: this.calendarIconRef,
-                    onClick: () => {
-                      const { focusedInput } = this.state;
-                      if (focusedInput) {
-                        this.setState({ focusedInput: undefined });
-                      }
-                    },
-                  })
-                : undefined
-            }
-            inputIconPosition="after"
             customArrowIcon="-"
-            showDefaultInputIcon={datepicker}
             onClose={this.onClose}
             numberOfMonths={2}
             minimumNights={0}
+            autoComplete="date"
             {...attributes}
           />
         </InputGroup>
@@ -568,8 +552,6 @@ AvDateRange.propTypes = {
   ranges: PropTypes.oneOfType([PropTypes.bool, PropTypes.array, PropTypes.object]),
   onPickerFocusChange: PropTypes.func,
   defaultValues: PropTypes.object,
-  calendarIcon: PropTypes.node,
-  datepicker: PropTypes.bool,
   autoSync: PropTypes.bool,
 };
 
@@ -577,8 +559,6 @@ AvDateRange.contextTypes = { FormCtrl: PropTypes.object.isRequired };
 
 AvDateRange.defaultProps = {
   type: 'text',
-  calendarIcon: <Icon name="calendar" />,
-  datepicker: true,
 };
 
 export default AvDateRange;