+++ packages/page-header/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.1.10](https://github.com/Availity/availity-react/compare/@availity/page-header@11.1.9...@availity/page-header@11.1.10) (2021-12-14)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## 11.1.9 (2021-12-09)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.1.8](https://github.com/Availity/availity-react/compare/@availity/page-header@11.1.7...@availity/page-header@11.1.8) (2021-12-03)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## 11.1.7 (2021-12-03)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.1.6](https://github.com/Availity/availity-react/compare/@availity/page-header@11.1.5...@availity/page-header@11.1.6) (2021-12-02)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.1.5](https://github.com/Availity/availity-react/compare/@availity/page-header@11.1.4...@availity/page-header@11.1.5) (2021-12-01)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.1.4](https://github.com/Availity/availity-react/compare/@availity/page-header@11.1.3...@availity/page-header@11.1.4) (2021-11-30)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
 ## [11.1.3](https://github.com/Availity/availity-react/compare/@availity/page-header@11.1.2...@availity/page-header@11.1.3) (2021-11-19)
 
 **Note:** Version bump only for package @availity/page-header